changed everthing I was asked to change

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robotSubSystems/intake/IntakeConstants.java
@@ -4,5 +4,5 @@ import com.acmerobotics.dashboard.config.Config;
 
 @Config
 public class IntakeConstants {
-public static float intakePower = 1f;
+public static double intakePower = 1f;
 }
